UICONSET-58 move foreign components to the 'temp' folder
