+++ src/components/index.js
@@ -1,4 +1,3 @@
 export { ConsortiumManagerHeader } from './ConsortiumManagerHeader';
 export { FindConsortiumMember } from './FindConsortiumMember';
-export { IfConsortiumPermission } from './IfConsortiumPermission';
 export { SwitchActiveAffiliation } from './SwitchActiveAffiliation';
