+++ src/hooks/index.js
@@ -1,4 +1,3 @@
-export { useConsortiumPermissions } from './useConsortiumPermissions';
 export { useCurrentConsortium } from './useCurrentConsortium';
 export { useTenantKy } from './useTenantKy';
 export { useUserAffiliations } from './useUserAffiliations';
