+++ src/temp/PermissionsAccordion/helpers/filtersConfig.js
