+++ src/temp/README.md
@@ -1,0 +1,5 @@
+# `temp` folder
+
+This folder contains copied components from another module (`ui-users`) and is intended for temporary storage of duplicated code. Subsequently, it must be deleted after the required components have been moved to the shared source.
+
+The folder stores the `<PermissionSetDetails>` and `<PermissionSetForm>` components and all their dependencies.
