+++ src/temp/index.js
@@ -1,2 +1,3 @@
 export { default as PermissionSetDetails } from './PermissionSetDetails';
 export { default as PermissionSetForm } from './PermissionSetForm';
+export * from './hooks';
