--- test/jest/__mock__/IfConsortiumPermission.mock.js
@@ -1,3 +1,0 @@
-jest.mock('../../../src/components/IfConsortiumPermission', () => ({
-  IfConsortiumPermission: jest.fn(({ children }) => <>{children}</>),
-}));
