+++ test/jest/__mock__/index.js
@@ -1,3 +1,2 @@
-import './IfConsortiumPermission.mock';
 import './reactFinalFormArrays.mock';
 import './reactFinalFormListeners.mock';
